requirements.txt
streamlit
python-docx
fitz
PyMuPDF
groq 
python-dotenv
langchain
phidata 
